requirements.txt
Django==3.0.4
djangorestframework==3.11.0
coverage~=4.5.3
pylint-django
django-filter==2.2.0
django-mysql==3.3.0
mysqlclient~=1.4.4